#8: Move Utils.ts to /utils directory
